scripts/run_demo.py
#!/usr/bin/env python3
"""
HAWKEYE Demo Script

Runs simulated drone inspection with real-time detection visualization.

Usage:
    python scripts/run_demo.py
    python scripts/run_demo.py checkpoint=outputs/checkpoints/best.pt
"""

import hydra
from omegaconf import DictConfig, OmegaConf
import logging
from pathlib import Path

log = logging.getLogger(__name__)


@hydra.main(version_base=None, config_path="../configs", config_name="config")
def main(cfg: DictConfig) -> None:
    """Main entry point for demo."""

    log.info("=" * 60)
    log.info("HAWKEYE Demo - Simulated Drone Inspection")
    log.info("=" * 60)

    # TODO: Implement demo
    # 1. Load trained model
    # 2. Load or generate warehouse scene
    # 3. Generate flight path
    # 4. For each waypoint:
    #    - Render RGB-D frame
    #    - Run inference
    #    - Track detections
    #    - Visualize with overlays
    # 5. Generate demo video

    log.info("\n[TODO] Demo not yet implemented")
    log.info("Complete training first, then implement demo visualization")


if __name__ == "__main__":
    main()
